refactor(api): move snowplow out of api
